Moved slider logic into its own hook (WIP)
